+++ librarian/sysu/string.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-int sysu_strlen(const char str[]);
+long long sysu_strlen(const char str[]);
 
 int sysu_strcmp(const char lhs[], const char rhs[]);
