+++ librarian/sysu/unistd.h
@@ -8,9 +8,9 @@ extern "C" {
 
 int sysu_close(int fd);
 
-int sysu_read(int fd, char buf[], int count);
+long long sysu_read(int fd, char buf[], long long count);
 
-int sysu_write(int fd, const char buf[], int count);
+long long sysu_write(int fd, const char buf[], long long count);
 
 int sysu_fork();
